Handle b<0

+++ include/fc/io/raw.hpp
@@ -172,7 +172,7 @@ namespace fc {
       uint64_t v = 0; char b = 0; uint8_t by = 0;
       do {
           s.get(b);
-          if( by >= 64 || (by == 63 && b > 1) )
+          if( by >= 64 || (by == 63 && uint8_t(b) > 1) )
              FC_THROW_EXCEPTION( overflow_exception, "Invalid packed unsigned_int!" );
           v |= uint64_t(uint8_t(b) & 0x7f) << by;
           by += 7;
